+++ src/main/java/org/apiphany/ApiClientFluentAdapter.java
@@ -7,7 +7,6 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.function.Supplier;
 
-import org.apiphany.RequestParameters.ParameterFunction;
 import org.apiphany.client.ExchangeClient;
 import org.apiphany.header.Headers;
 import org.apiphany.lang.collections.Maps;
@@ -43,7 +42,7 @@ public class ApiClientFluentAdapter extends ApiRequest<Object> {
 	 * @param apiClient the underlying API client
 	 */
 	protected ApiClientFluentAdapter(final ApiClient apiClient) {
-		this.apiClient = apiClient;
+		this.apiClient = Objects.requireNonNull(apiClient);
 	}
 
 	/**
@@ -274,7 +273,7 @@ public class ApiClientFluentAdapter extends ApiRequest<Object> {
 
 	/**
 	 * Sets the request parameters. If the input request parameters is {@code null} then the request
-	 * parameters will be set to a an empty map.
+	 * parameters will be set to an empty map.
 	 *
 	 * @param requestParams request parameters
 	 * @return this
@@ -385,7 +384,7 @@ public class ApiClientFluentAdapter extends ApiRequest<Object> {
 	 * @param stream value to set
 	 * @return this
 	 */
-	public ApiClientFluentAdapter stream(boolean stream) {
+	public ApiClientFluentAdapter stream(final boolean stream) {
 		this.stream = stream;
 		return this;
 	}
@@ -406,7 +405,7 @@ public class ApiClientFluentAdapter extends ApiRequest<Object> {
 	 * @return this
 	 */
 	public ApiClientFluentAdapter meters(final String prefix) {
-		return meters(BasicMeters.of(prefix));
+		return meters(BasicMeters.of(apiClient.getMeterRegistry(), prefix));
 	}
 
 	/**
@@ -417,7 +416,7 @@ public class ApiClientFluentAdapter extends ApiRequest<Object> {
 	 * @return this
 	 */
 	public ApiClientFluentAdapter meters(final String prefix, final Tags tags) {
-		return meters(BasicMeters.of(prefix, tags));
+		return meters(BasicMeters.of(apiClient.getMeterRegistry(), prefix, tags));
 	}
 
 	/**
@@ -427,7 +426,7 @@ public class ApiClientFluentAdapter extends ApiRequest<Object> {
 	 * @return this
 	 */
 	public ApiClientFluentAdapter metersOnMethod(final String prefix) {
-		return meters(BasicMeters.onCallerMethod(prefix));
+		return meters(BasicMeters.onCallerMethod(apiClient.getMeterRegistry(), prefix));
 	}
 
 	/**
@@ -438,7 +437,7 @@ public class ApiClientFluentAdapter extends ApiRequest<Object> {
 	 * @return this
 	 */
 	public ApiClientFluentAdapter metersOnMethod(final String prefix, final Tags tags) {
-		return meters(BasicMeters.onCallerMethod(prefix, tags));
+		return meters(BasicMeters.onCallerMethod(apiClient.getMeterRegistry(), prefix, tags));
 	}
 
 	/**