+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ---
 
+`1.0.1`
+
+- Extracted `ParameterFunction` inner class from `RequestParameters` class.
+- Upgraded parent to `3.5.0` and other dependencies defined in it.
+- Added `MeterRegistry` to `ApiClient`.
+- Added `MeterRegistry` to most methods in `BasicMeters`.
+- Changed `ClientProperties` to use `Duration` instead of `int` (milliseconds).
+
+---
+
 `1.0.0`
 
 - First release.